+++ src/getStepStartStates.test.ts
@@ -74,6 +74,7 @@ import {
   STEP_RM_CDN_PROFILE,
   STEP_RM_CDN_ENDPOINTS,
 } from './steps/resource-manager/cdn';
+import { STEP_RM_BATCH_ACCOUNT } from './steps/resource-manager/batch';
 
 describe('getStepStartStates', () => {
   test('all steps represented', () => {
@@ -136,6 +137,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_SERVICE_BUS_SUBSCRIPTIONS]: { disabled: true },
       [STEP_RM_CDN_PROFILE]: { disabled: true },
       [STEP_RM_CDN_ENDPOINTS]: { disabled: true },
+      [STEP_RM_BATCH_ACCOUNT]: { disabled: true },
     });
   });
 
@@ -194,6 +196,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_SERVICE_BUS_SUBSCRIPTIONS]: { disabled: true },
       [STEP_RM_CDN_PROFILE]: { disabled: true },
       [STEP_RM_CDN_ENDPOINTS]: { disabled: true },
+      [STEP_RM_BATCH_ACCOUNT]: { disabled: true },
     });
   });
 
@@ -252,6 +255,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_SERVICE_BUS_SUBSCRIPTIONS]: { disabled: false },
       [STEP_RM_CDN_PROFILE]: { disabled: false },
       [STEP_RM_CDN_ENDPOINTS]: { disabled: false },
+      [STEP_RM_BATCH_ACCOUNT]: { disabled: false },
     });
   });
 });